⚡feat : People 출력 로직 분리
