+++ src/test/java/ladder/util/LadderOutputConverterTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.List;
 import java.util.stream.Stream;
@@ -34,8 +35,15 @@ public class LadderOutputConverterTest {
     @DisplayName("5자 미만일 경우 공백을 채워 5자로 출력한다.")
     @Test
     void result_output() {
-        Results results = new Results("꽝","5000","3000");
+        Results results = new Results("꽝", "5000", "3000");
 
         assertThat(LadderOutputConverter.resultOutput(results)).isEqualTo("    꽝  5000  3000");
+    }
+
+
+    @ParameterizedTest(name = "사람 이름이 5자 미만일 경우 공백을 채워 5자로 출력한다.")
+    @ValueSource(strings = {"a", "ab", "abc", "abcd", "abcde"})
+    void personname_output(String text) {
+        assertThat(LadderOutputConverter.outputPersonName(text).length()).isEqualTo(5);
     }
 }
